hai ca am facut o si pe asta

+++ Sterge.cs
@@ -1,4 +1,4 @@
-using System;
+﻿using System;
 using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data;
@@ -26,9 +26,59 @@ namespace b2
         {
             Forms.w2.Show();
         }
+        int ok(string cnp)
+        {
+            string telefon = "";
+            using (OleDbConnection con = new OleDbConnection(Conexiune.path))
+            {
+                string query = "SELECT NumarTelefon FROM ELEV WHERE CNP = @a";
+                con.Open();
+
+                using (OleDbCommand com = new OleDbCommand(query, con))
+                {
+                    com.Parameters.AddWithValue("@a", cnp);
+
+                    using (OleDbDataReader reader = com.ExecuteReader())
+                    {
+                        if (reader.Read())
+                        {
+                            telefon = reader["NumarTelefon"].ToString();
+                        }
+                        else
+                        {
+                            MessageBox.Show("CNP inexistent", "Eroare", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                            return 1;
+                        }
+                    }
+                }
+            }
+
+            using (OleDbConnection con = new OleDbConnection(Conexiune.path))
+            {
+                string query = @"SELECT COUNT(*) FROM Imprumuturi WHERE NumarTelefon = @a";
+                con.Open();
+
+                using (OleDbCommand com = new OleDbCommand(query, con))
+                {
+                    com.Parameters.AddWithValue(@"a", telefon);
+                    int chestie = (int)com.ExecuteScalar();
+                    if(chestie > 0)
+                    {
+                        MessageBox.Show("Nu au fost returnate toate cărțile", "Eroare", MessageBoxButtons.OK, MessageBoxIcon.Error);
+                        return 2;
+                    }
+                }
+            }
+            return 3;
+        }
         private void button1_Click(object sender, EventArgs e)
         {
             string cnp = textBox1.Text;
+            int ceva = ok(cnp);
+            if(ceva == 1 || ceva == 2)
+            {
+                return;
+            }
             using (OleDbConnection con = new OleDbConnection(Conexiune.path))
             {
                 string query = @"DELETE FROM Elev WHERE CNP = @a";
